style: move UI-related code to :ui
