+++ build.gradle
@@ -1,15 +1,24 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
+    apply from: 'plugins.gradle'
+    apply from: 'implementations.gradle'
+    apply from: 'versions.gradle'
+
     repositories {
         google()
         mavenCentral()
     }
-    dependencies {
-        classpath "com.android.tools.build:gradle:7.0.3"
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.0"
 
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
+    dependencies {
+        classpath _plugins.gradle
+        classpath _plugins.kotlin
+    }
+}
+
+allprojects {
+    repositories {
+        google()
+        maven { url 'https://jitpack.io' }
+        mavenCentral()
     }
 }
