+++ plugins.gradle
@@ -1,0 +1,10 @@
+buildscript {
+    ext {
+        apply from: 'versions.gradle'
+
+        _plugins = [
+            gradle: "com.android.tools.build:gradle:${versions.gradle}",
+            kotlin: "org.jetbrains.kotlin:kotlin-gradle-plugin:${versions.kotlin}"
+        ]
+    }
+}
