+++ ui/build.gradle
@@ -1,19 +1,19 @@
 plugins {
-    id 'com.android.application'
+    id 'com.android.library'
     id 'kotlin-android'
 }
 
 android {
-    compileSdk versions.ingresso.sdk.compile
+    compileSdk 31
 
     defaultConfig {
-        applicationId 'com.jeanbarrossilva.ingresso'
         minSdk versions.ingresso.sdk.min
         targetSdk versions.ingresso.sdk.target
         versionCode versions.ingresso.code
         versionName versions.ingresso.name
 
         testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
+        consumerProguardFiles 'consumer-rules.pro'
     }
 
     buildTypes {
@@ -34,5 +34,6 @@ android {
 }
 
 dependencies {
-    implementation implementations.ingresso.ui
+    implementation implementations.appcompat
+    implementation implementations.material
 }
