+++ versions.gradle
@@ -1,0 +1,20 @@
+buildscript {
+    ext {
+        versions = [
+            appcompat: '1.4.0',
+            gradle: '7.0.3',
+            ingresso: [
+                code: 1,
+                name: '1.0.0',
+                sdk: [
+                    compile: 31,
+                    min: 21,
+                    target: 31
+                ]
+            ],
+            java: JavaVersion.VERSION_1_8,
+            kotlin: '1.6.0',
+            material: '1.4.0'
+        ]
+    }
+}
